Extract common method from knight and king
Knight and king doesn't follow specific patterns like Bishop and Rook,
where we keep calculating moves in a particular direction unless we
reach out of the board or get interrupted by a piece.

Instead they have few countable valid moves. I didn't try to abstract
this because it is very likely that a knight will never change the way
it moves. So the way we calculate moves will never change. Hence, it
doesn't need to be abstracted.