Move possible pieces selection to Game
Possible pieces which we can select to make a move depends on the Game
state. No. of possible pieces you can select will depend if you're in
check or not.

Check is a game rule so anything that depends on it should belong to
game. That's why I'm moving possible pieces selection functionality here