Only select pieces which gets us out of check
When you're in check your pieces selections gets limited. The piece
moves get limited as well.

The hard part in implementing this was that I was only considering only
"escaping" as a way to get out of the check. But we can capture the
checking piece to get out of check too.