Add test for queen_spec.rb
I deleted it earlier because I thought that queen is made up of rook &
bishop. We have tested both of them, so queen should work fine if rook &
bishop tests are green.

I added this now because I need to know that queen is working fine, so I
can focus on limiting pieces movement when they are in check by Queen