Fix: don't let pawn capture its own pieces
Earlier when the king was placed in front of its own pawn. The capturing
of own king was a legal move for pawn.

I found this when black king was on e6 and brown queen was on f6. The e5
pawn was capturing the e6 king. Hence, when we were checking if black
king is in check during our calculation of legal moves then there was no
king.

Therefore, king.position was failing because board.king(color) was
returning an array of pieces because the method couldn't find king.