Replacing difference with sameness in castling
Both queenside and kingside castling difference are replaced by sameness
This will help in finding a nice abstraction for castling?.

How did I reach to this?
Both queenside castling and kingside castling are in face castling. So
if castling is a bigger abstraction then there must be smaller
abstraction within it to support queenside and kingside castling.